Don't negotiate zero capabilities.
Setting the requested capabilities to nil or an empty slice will result
in a hang (at least on solanum) because we send `CAP REQ :`, which is
not answered by any ACK or otherwise.

In the future, it may be a good idea to send the `CAP LS 302` anyway,
but then immediately send a `CAP END`. Allowing us to be informed of new
or removed capabilities. However as we have none to request anyway....
$shrug.